requirements.txt
# pip install "django>=2.1,<2.2"
# 2.2.11
django
wheel
pylint

django-split-settings
django-crispy-forms
django-tables2
django-filter

# python-decouple
# django-ckeditor
# django-mathfilters
# python-social-auth
# social-auth-app-django
# django-debug-toolbar
# django-crispy-forms
# django-templated-email
# django-next-prev
# fdb
# django-firebird==1.11
# # instalamos manualmente los driver 2.0 de firebird

# # -------------------------------------------------
# # dentro de pruebas
# # -------------------------------------------------
# django-widget-tweaks
# django-tables2
# django-search-views
# django-filter
# django-bootstrap3
